[NuGet] Images now loaded using configured http message handler
Images were being loaded by the HttpClientHandler and not the
configured HttpMessageHandler used by the IDE. Also when checking
package sources the HttpClientHandler. Now the creation of the
HttpClient used to load images and check package sources are
available uses the MonoDevelopHttpHandlerResourceV3Provider which
uses a native http message handler is used if configured.